Bump version
this is just a re-release of the old version because for unknown reasons
our devpi had tar file of the previous version that was missing assets.

a new dist build (without changes to the manifest or otherwise) produced
a valid and complete package, which was then uploaded with this version
number.